Add option to set workerPoolConcurrency to zero
