+++ packages/plugin-git/sources/GitFetcher.ts
@@ -67,7 +67,7 @@ export class GitFetcher implements Fetcher {
         compressionLevel: opts.project.configuration.get(`compressionLevel`),
         prefixPath: structUtils.getIdentVendorPath(locator),
         stripComponents: 1,
-        poolSize: opts.project.configuration.getLimit(`workerPoolConcurrency`),
+        poolSize: opts.project.configuration.get(`workerPoolConcurrency`),
       });
     });
   }
