+++ packages/plugin-github/sources/GithubFetcher.ts
@@ -64,7 +64,7 @@ export class GithubFetcher implements Fetcher {
         compressionLevel: opts.project.configuration.get(`compressionLevel`),
         prefixPath: structUtils.getIdentVendorPath(locator),
         stripComponents: 1,
-        poolSize: opts.project.configuration.getLimit(`workerPoolConcurrency`),
+        poolSize: opts.project.configuration.get(`workerPoolConcurrency`),
       });
     });
   }
