+++ packages/plugin-http/sources/TarballHttpFetcher.ts
@@ -46,7 +46,7 @@ export class TarballHttpFetcher implements Fetcher {
       compressionLevel: opts.project.configuration.get(`compressionLevel`),
       prefixPath: structUtils.getIdentVendorPath(locator),
       stripComponents: 1,
-      poolSize: opts.project.configuration.getLimit(`workerPoolConcurrency`),
+      poolSize: opts.project.configuration.get(`workerPoolConcurrency`),
     });
   }
 }
