+++ packages/plugin-npm/sources/NpmHttpFetcher.ts
@@ -58,7 +58,7 @@ export class NpmHttpFetcher implements Fetcher {
       compressionLevel: opts.project.configuration.get(`compressionLevel`),
       prefixPath: structUtils.getIdentVendorPath(locator),
       stripComponents: 1,
-      poolSize: opts.project.configuration.getLimit(`workerPoolConcurrency`),
+      poolSize: opts.project.configuration.get(`workerPoolConcurrency`),
     });
   }
 }
