+++ packages/plugin-npm/sources/NpmSemverFetcher.ts
@@ -68,7 +68,7 @@ export class NpmSemverFetcher implements Fetcher {
       compressionLevel: opts.project.configuration.get(`compressionLevel`),
       prefixPath: structUtils.getIdentVendorPath(locator),
       stripComponents: 1,
-      poolSize: opts.project.configuration.getLimit(`workerPoolConcurrency`),
+      poolSize: opts.project.configuration.get(`workerPoolConcurrency`),
     });
   }
