+++ packages/yarnpkg-core/sources/worker-zip/index.d.ts
@@ -1,2 +1,1 @@
 export function getContent(): string;
-export type {ConvertToZipPayload} from './Worker';
