+++ packages/yarnpkg-core/tests/tgzUtils.test.ts
@@ -1,5 +1,4 @@
 import {npath, xfs}  from '@yarnpkg/fslib';
-import pLimit        from 'p-limit';
 
 import * as tgzUtils from '../sources/tgzUtils';
 
@@ -11,7 +10,7 @@ describe(`tgzUtils`, () => {
           npath.join(__dirname, `fixtures/carbon-icons-svelte-10.21.0.tgz`),
         ),
       );
-      await expect(tgzUtils.convertToZip(data, {compressionLevel: 0, poolSize: pLimit(2)})).resolves.toBeTruthy();
+      await expect(tgzUtils.convertToZip(data, {compressionLevel: 0, poolSize: 2})).resolves.toBeTruthy();
     });
 
     it(`should be able to convert a tgz without compression`, async () => {
@@ -20,7 +19,16 @@ describe(`tgzUtils`, () => {
           npath.join(__dirname, `fixtures/parse5-0.0.28.tgz`),
         ),
       );
-      await expect(tgzUtils.convertToZip(data, {compressionLevel: 0, poolSize: pLimit(2)})).resolves.toBeTruthy();
+      await expect(tgzUtils.convertToZip(data, {compressionLevel: 0, poolSize: 2})).resolves.toBeTruthy();
+    });
+
+    it(`should be able to convert a tgz without any workers`, async () => {
+      const data = await xfs.readFilePromise(
+        npath.toPortablePath(
+          npath.join(__dirname, `fixtures/parse5-0.0.28.tgz`),
+        ),
+      );
+      await expect(tgzUtils.convertToZip(data, {compressionLevel: 0, poolSize: 0})).resolves.toBeTruthy();
     });
   });
 });
